Duplicated Transifex to separate pull and push
This is a safety measure to avoid keeping modifying the tx.cmd script, commenting and decommenting the same script all the time based on what we need to do.
Now the 2 scripts perform 2 different actions:

tx_pull.cmd is to be used when we need to download the last translations made on Tx website.
tx_push.cmd is to be used when the game got updated and therefore there might be new or different strings to translate in the new_language template, so we need to push the new strings to Transifex.