Pulling latest translations from Transifex.
Some strings that were previously translated were replaced by its english version because something changes since the last template update. If anybody wants to fix the strings, they just need to go on the appropriate Transifex project and restore the string using the translation history on the right (or update the translation to reflect the changes, if the previous translation is no longer 100% correct).
There are also new strings that will certainly need to be translated from scratch, but they're not many.